# src/main/java/org/example/cuidadodemascotas/servicemicroservice/apis/controller/PaymentMethodApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.2.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package org.example.cuidadodemascotas.servicemicroservice.apis.controller;

import org.example.cuidadodemascotas.servicemicroservice.apis.dto.PaymentMethodRequestDTO;
import org.example.cuidadodemascotas.servicemicroservice.apis.dto.PaymentMethodResponseDTO;
import org.example.cuidadodemascotas.servicemicroservice.apis.dto.PaymentMethodResult;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-11-20T03:45:32.928054400-03:00[America/Asuncion]")
@Validated
@Tag(name = "PaymentMethod", description = "the PaymentMethod API")
public interface PaymentMethodApi {

    /**
     * POST /payment-methods : Crear un nuevo método de pago
     *
     * @param paymentMethodRequestDTO  (required)
     * @return Método de pago creado exitosamente (status code 201)
     *         or Solicitud inválida (status code 400)
     *         or Error interno del servidor (status code 500)
     */
    @Operation(
        operationId = "createPaymentMethod",
        summary = "Crear un nuevo método de pago",
        tags = { "PaymentMethod" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Método de pago creado exitosamente", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = PaymentMethodResponseDTO.class))
            }),
            @ApiResponse(responseCode = "400", description = "Solicitud inválida"),
            @ApiResponse(responseCode = "500", description = "Error interno del servidor")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/payment-methods",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<PaymentMethodResponseDTO> createPaymentMethod(
        @Parameter(name = "PaymentMethodRequestDTO", description = "", required = true) @Valid @RequestBody PaymentMethodRequestDTO paymentMethodRequestDTO
    );


    /**
     * GET /payment-methods : Obtener todos los métodos de pago
     *
     * @param page  (optional, default to 0)
     * @param size  (optional, default to 20)
     * @return Métodos de pago obtenidos exitosamente (status code 200)
     *         or Error interno del servidor (status code 500)
     */
    @Operation(
        operationId = "getPaymentMethods",
        summary = "Obtener todos los métodos de pago",
        tags = { "PaymentMethod" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Métodos de pago obtenidos exitosamente", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = PaymentMethodResult.class))
            }),
            @ApiResponse(responseCode = "500", description = "Error interno del servidor")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/payment-methods",
        produces = { "application/json" }
    )
    
    ResponseEntity<PaymentMethodResult> getPaymentMethods(
        @Parameter(name = "page", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "page", required = false, defaultValue = "0") Integer page,
        @Parameter(name = "size", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "size", required = false, defaultValue = "20") Integer size
    );

}