# src/main/resources/config-openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Invoice and Payments Microservice API
  description: Microservicio para gestión de facturas y pagos
  version: 1.0.0
servers:
  - url: http://localhost:8082
    description: Servicio de Facturas y Pagos
paths:
...
  /payment-methods:
    get:
      tags:
        - PaymentMethod
      summary: Obtener todos los métodos de pago
      operationId: getPaymentMethods
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Métodos de pago obtenidos exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentMethodResult'
        '500':
          description: Error interno del servidor

    post:
      tags:
        - PaymentMethod
      summary: Crear un nuevo método de pago
      operationId: createPaymentMethod
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentMethodRequestDTO'
      responses:
        '201':
          description: Método de pago creado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentMethodResponseDTO'
        '400':
          description: Solicitud inválida
        '500':
          description: Error interno del servidor

  # ==================== PAYMENT DETAILS ====================
...
components:
  schemas:
...
    PaymentMethodRequestDTO:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          maxLength: 60
    PaymentMethodResponseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    PaymentMethodResult:
      type: object
      properties:
        totalCount:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/PaymentMethodResponseDTO'

    # ==================== PAYMENT DETAIL ====================
...